some small changes

+++ Laboratorium6/Laboratorium6/Vector.h
@@ -17,7 +17,7 @@ public:
 	Vector& operator+(const Vector&);
 	Vector& operator-(const Vector&);
 	Vector& operator*(double);
-	float operator!();
+	double operator!();
 	float operator%(Vector&);
